Added Dissertation to Github
DISCONTINUED: This artefact is no longer recieving attention from me, as what I did here is being used in another project: Planet_Simulation.
This is here in case I need to reference back to it at anytime, as well as to see how far I have come as a programmer.